+++ src/navigator/OccupancyGridClientNav.js
@@ -14,6 +14,7 @@
  *   * serverName (optional) - the action server name to use for navigation, like '/move_base'
  *   * actionName (optional) - the navigation action name, like 'move_base_msgs/MoveBaseAction'
  *   * rootObject (optional) - the root object to add the click listeners to and render robot markers to
+ *   * withOrientation (optional) - if the Navigator should consider the robot orientation (default: false)
  *   * viewer - the main viewer to render to
  */
 NAV2D.OccupancyGridClientNav = function(options) {
@@ -26,6 +27,7 @@ NAV2D.OccupancyGridClientNav = function(options) {
   this.actionName = options.actionName || 'move_base_msgs/MoveBaseAction';
   this.rootObject = options.rootObject || new createjs.Container();
   this.viewer = options.viewer;
+  this.withOrientation = options.withOrientation || false;
 
   this.navigator = null;
 
@@ -33,17 +35,21 @@ NAV2D.OccupancyGridClientNav = function(options) {
   var client = new ROS2D.OccupancyGridClient({
     ros : this.ros,
     rootObject : this.rootObject,
-    continuous : continuous
+    continuous : continuous,
+    topic : topic
   });
   client.on('change', function() {
     that.navigator = new NAV2D.Navigator({
       ros : that.ros,
       serverName : that.serverName,
       actionName : that.actionName,
-      rootObject : that.rootObject
+      rootObject : that.rootObject,
+      withOrientation : that.withOrientation
     });
     
     // scale the viewer to fit the map
+    that.viewer.shift(client.currentGrid.pose.position.x, client.currentGrid.pose.position.y);
     that.viewer.scaleToDimensions(client.currentGrid.width, client.currentGrid.height);
+    that.viewer.shift(client.currentGrid.x,client.currentGrid.y);
   });
 };