+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+2013-07-08 - **r2**
+ * Shift added to grid client to center on map [(rctoris)](https://github.com/rctoris/)
+ * OccupancyGridClientNav now correctly uses the topic parameter [(rctoris)](https://github.com/rctoris/)
+ * Navigator now allows click and hold for orientation [(kunzel)](https://github.com/kunzel/)
+
 2013-04-15 - **r1**
  * Initial development of NAV2D [(rctoris)](https://github.com/rctoris/)